requirements.txt
alembic==1.4.1
aniso8601==9.0.1
argcomplete==1.12.3
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
azure-core==1.17.0
azure-storage-blob==12.8.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
boto3==1.18.21
botocore==1.21.21
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
cloudpickle==1.6.0
coverage==5.5
cryptography==3.4.7
databricks-cli==0.15.0
debugpy==1.4.1
decorator==5.0.9
dill==0.3.4
distlib==0.3.2
docker==5.0.0
dotmap==1.3.24
elastic-apm==6.3.3
elasticsearch==7.10.0
entrypoints==0.3
faiss-cpu==1.7.1.post2
farm==0.8.0
farm-haystack==0.9.0
fastapi==0.68.0
filelock==3.0.12
Flask==1.1.4
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flask-restplus==0.13.0
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.1
grpcio==1.37.1
grpcio-tools==1.37.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.3.0
huggingface-hub==0.0.8
idna==2.10
importlib-metadata==4.6.4
ipykernel==6.1.0
ipython==7.26.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
langdetect==1.0.9
lxml==4.6.3
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mlflow==1.13.1
mmh3==3.0.0
more-itertools==8.8.0
msrest==0.6.21
networkx==2.6.2
nltk==3.6.2
numpy==1.21.1
oauthlib==3.1.1
packaging==21.0
pandas==1.3.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.2.0
pluggy==0.13.1
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.8.0
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pymilvus==1.1.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
querystring-parser==1.2.4
rdflib==6.0.0
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.16.0
sklearn==0.0
smmap==4.0.0
SPARQLWrapper==1.8.5
SQLAlchemy==1.4.22
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.1
starlette==0.14.2
tabulate==0.8.9
threadpoolctl==2.2.0
tika==1.24
tokenizers==0.10.3
toml==0.10.2
torch==1.8.1
torchvision==0.10.0
tornado==6.1
tox==3.24.1
tqdm==4.61.2
traitlets==5.0.5
transformers==4.6.1
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.14.0
validators==0.18.2
virtualenv==20.7.2
wcwidth==0.2.5
weaviate-client==3.0.0
websocket-client==1.2.1
Werkzeug==0.16.1
zipp==3.5.0
